+++ stac_fastapi/opensearch/stac_fastapi/opensearch/database_logic.py
@@ -21,7 +21,7 @@ from stac_fastapi.opensearch.config import (
 )
 from stac_fastapi.opensearch.config import OpensearchSettings as SyncSearchSettings
 from stac_fastapi.sfeos_helpers import filter
-from stac_fastapi.sfeos_helpers.database_logic_helpers import (
+from stac_fastapi.sfeos_helpers.database import (
     apply_free_text_filter_shared,
     apply_intersects_filter_shared,
     create_index_templates_shared,
